requirements.txt
pafy
pydub
python-telegram-bot
bs4
requests
youtube-dl
